cloudrun_train_jcl/requirements.txt
Flask
google-cloud-storage
google-cloud-bigquery
pandas
numpy
scikit-learn
joblib